Centralize configuration
With this change, two files are used for configuration:

`config.dist.php` contains default settings:

- the `clubId` = 198 (i.e. HGDF)
- flying `sites` = ['Neustadt-Glewe', 'Hörpel', 'Altenmedingen'] (for HGDF)
- `db` credentials (`servername`, `username`, `password`, `dbname`, `port`)
- the known clubs

DB credentials are read from environment variables, if set:

- `DATABASE_HOST`
- `DATABASE_USER`
- `DATABASE_PASSWORD`
- `DATABASE_NAME`
- `DATABASE_PORT`

If those are not given, then

- `localhost` is used for the server,
- `flugplanung` for the database name,
- `3306` for the port

An optional file `config.php` is read and overrides the settings from
`config.dist.php`. It is ignored in `.gitingore` and can be used to
customize settings for a specific setup.